+++ src/lib.rs
@@ -14,6 +14,7 @@ mod alignment;
 mod float_ext;
 mod length;
 mod mask;
+mod quaternion;
 mod scalar;
 mod vector;
 pub use aliases::*;
@@ -21,6 +22,7 @@ pub use alignment::*;
 pub use float_ext::*;
 pub use length::*;
 pub use mask::*;
+pub use quaternion::*;
 pub use scalar::*;
 pub use vector::*;