+++ tests/integrations/mint.rs
@@ -1,7 +1,7 @@
 use assert_impl_trait::assert_impl;
 use ggmath::{
-    Alignment, Mask, Mask2, Mask2U, Mask3, Mask3U, Mask4, Mask4U, Scalar, Vec2, Vec2U, Vec3, Vec3U,
-    Vec4, Vec4U, Vector, vec2, vec3, vec4,
+    Alignment, Mask, Mask2, Mask2U, Mask3, Mask3U, Mask4, Mask4U, Quat, QuatU, Quaternion, Scalar,
+    Vec2, Vec2U, Vec3, Vec3U, Vec4, Vec4U, Vector, vec2, vec3, vec4,
 };
 use mint::IntoMint;
 
@@ -22,6 +22,10 @@ assert_impl!(
         Vector<4, T, A>: IntoMint<MintType = mint::Vector4<T>>,
         Vector<4, T, A>: From<mint::Vector4<T>>,
         Vector<4, T, A>: Into<mint::Vector4<T>>,
+
+        Quaternion<T, A>: IntoMint<MintType = mint::Quaternion<T>>,
+        Quaternion<T, A>: From<mint::Quaternion<T>>,
+        Quaternion<T, A>: Into<mint::Quaternion<T>>,
 
         Mask<2, T, A>: IntoMint<MintType = mint::Vector2<bool>>,
         Mask<2, T, A>: From<mint::Vector2<bool>>,
@@ -60,6 +64,12 @@ fn vector() {
 
     let val: Vec4U<i32> = vec4!(1, 2, 3, 4);
     assert_eq!(val, mint::Vector4::from(val).into());
+
+    let val: Quat<i32> = Quat::new(1, 2, 3, 4);
+    assert_eq!(val, mint::Quaternion::from(val).into());
+
+    let val: QuatU<i32> = QuatU::new(1, 2, 3, 4);
+    assert_eq!(val, mint::Quaternion::from(val).into());
 
     let val: Mask2<i32> = Mask2::new(false, true);
     assert_eq!(val, mint::Vector2::<bool>::from(val).into());